Added basic inventory

+++ src/entities/mushrooms/BrownMushroom.java
@@ -14,7 +14,10 @@ public class BrownMushroom extends Collectable{
 
     @Override
     public void collect() {
-        World.entityManager.removeEntity(this);
+        boolean b = World.player.inv.addItem(this);
+        if(b){
+            World.entityManager.removeEntity(this);
+        }
     }
 
     @Override
